rename OrgContentController to OrgResourceController
